Move the shared test harness into `AbstractTensorTest` (wala/ML#635).
The five `test(...)` overloads, `containsPrimitiveByFieldWalk`,
`summarizeContext`, `getFunctionSignatures`, `getPathFiles`,
`CONTEXT_METHOD_PATTERN`, `CALL_GRAPH_DUMP_NODE_LIMIT`, and the
`LOGGER` move from `TestTensorflow2Model` to the base so the upcoming
feature-area carve classes need only extend it. Members the subclass
calls become `protected`; the `LOGGER` renames to the base's class,
staying under the package prefix the volume gate listens on.
